% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdpsd.R
\name{dfa2fd}
\alias{dfa2fd}
\title{DFA slope (H in DFA)}
\usage{
dfa2fd(sap)
}
\arguments{
\item{sap}{Self-afinity parameter estimate based on DFA slope (e.g., \code{\link{fd.sda}})).}
}
\value{
An informed estimate of the Fractal Dimension, see Hasselman(2013) for details.
}
\description{
Conversion formula: Detrended Fluctuation Analysis (DFA) estimate of the Hurst exponent (a self-affinity parameter \code{sap}) to an informed estimate of the (fractal) dimension (FD).
}
\details{
The DFA slope (H) will be converted to a dimension estimate using:

\deqn{D_{DFA}\approx 2-(\tanh(\log(3)*sap)) }{D_{DFA} ≈ 2-(tanh(log(3)*sap)) }
}
\examples{
# Informed FD of white noise
dfa2fd(0.5)

# Informed FD of Pink noise
dfa2fd(1)

# Informed FD of blue noise
dfa2fd(0.1)
}
\author{
Fred Hasselman
}
\references{
Hasselman, F. (2013). When the blind curve is finite: dimension estimation and model inference based on empirical waveforms. Frontiers in Physiology, 4, 75. \url{http://doi.org/10.3389/fphys.2013.00075}
}

