% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdpsd.R
\name{fd.sda}
\alias{fd.sda}
\title{Standardised Dispersion Analysis (SDA).}
\usage{
fd.sda(y, fs = NULL, normalize = TRUE, dtrend = FALSE,
  scales = dispersion(y)$scale, fitRange = c(scales[2],
  scales[length(scales) - 1]), plot = FALSE)
}
\arguments{
\item{normalize}{}
}
\value{
Slope and FD estimate based on Hasselman(2013)
}
\description{
fd.sda
}
\author{
Fred Hasselman
}
\references{
Hasselman, F. (2013). When the blind curve is finite: dimension estimation and model inference based on empirical waveforms. Frontiers in Physiology, 4, 75. \url{http://doi.org/10.3389/fphys.2013.00075}
}
\seealso{
Other FD estimators: \code{\link{fd.dfa}},
  \code{\link{fd.psd}}
}

