% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCS-TB.R
\name{growth.ac.cond}
\alias{growth.ac.cond}
\title{Conditional Autocatlytic Growth: Iterating differential equations (maps)}
\usage{
growth.ac.cond(Y0 = 0.01, r = 0.1, k = 2, cond = cbind.data.frame(Y =
  0.2, par = "r", val = 2), N = 100)
}
\arguments{
\item{Y0}{Initial value}

\item{r}{Growth rate parameter}

\item{k}{Carrying capacity}

\item{cond}{Conditional rules passed as a data.frame of the form: cbind.data.frame(Y = ..., par = ..., val = ...)}

\item{N}{Length of the time series}
}
\description{
Conditional Autocatlytic Growth: Iterating differential equations (maps)
}
\examples{
# Plot with the default settings
xyplot(growth.ac.cond())

# The function such that it can take a set of conditional rules and apply them sequentially during the iterations.
# The conditional rules are passed as a `data.frame`
(cond <- cbind.data.frame(Y = c(0.2, 0.6), par = c("r", "r"), val = c(0.5, 0.1)))
xyplot(growth.ac.cond(cond=cond))

# Combine a change of `r` and a change of `k`
(cond <- cbind.data.frame(Y = c(0.2, 1.99), par = c("r", "k"), val = c(0.5, 3)))
xyplot(growth.ac.cond(cond=cond))

# A fantasy growth process
(cond <- cbind.data.frame(Y = c(0.1, 1.99, 1.999, 2.5, 2.9), par = c("r", "k", "r", "r","k"), val = c(0.3, 3, 0.9, 0.1, 1.3)))
xyplot(growth.ac.cond(cond=cond))
}
\author{
Fred Hasselman
}
\seealso{
\code{\link{growth.ac.cond}}

Other autocatalytic growth functions: \code{\link{growth.ac}}
}

