% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inIT.R
\name{in.IT}
\alias{in.IT}
\title{Initialise It}
\usage{
in.IT(need = NULL, inT = TRUE)
}
\arguments{
\item{need}{A vector of package names to be loaded. The wrapper functions have a predefinded \code{need} list and can be used as shortcuts (see details).}

\item{inT}{Logical. If \code{TRUE} (default), packages in \code{need} wil be installed if they are not available on the system.}
}
\description{
Load and/or install R packages
}
\details{
\code{in.IT} will check if the Packages in the list argument \code{need} are installed on the system and load them. If \code{inT=TRUE} (default), it will first install the packages if they are not present and then proceed to load them.

The wrapper functions have a predefined need list:

\itemize{
\item \strong{in.IO}: Load I/O and data handling tools, \code{need = c("plyr","reshape2","RCurl","httr","dplyr","rio")}
\item \strong{in.PLOT}: Load tools for plotting \code{need = c("lattice","latticeExtra","gplots","ggplot2","grid","gridExtra","scales","effects","RColorBrewer")}
\item \strong{in.NLTS}: Load tools for Nonlinear Time Series Analysis, \code{need = c("fractaldim","fractalrock","RTisean","tsDyn","tseries","tseriesChaos")}
\item \strong{in.SN}: Load tools for signal analysis (Matlab style) \code{need = c("pracma","signal","EMD","hht","matlab","oce")}
}
}
\examples{
in.IT(c("reshape2", "plyr", "dplyr"))
}
\author{
Fred Hasselman
}
\seealso{
Other initialise packages: \code{\link{un.IT}}
}

