% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdpsd.R
\name{psd2fd}
\alias{psd2fd}
\title{psd2fd}
\usage{
psd2fd(sap)
}
\arguments{
\item{sap}{Self-afinity parameter estimate based on PSD slope (e.g., \code{\link{fd.psd}})).}
}
\value{
An informed estimate of the Fractal Dimension, see Hasselman(2013) for details.
}
\description{
Conversion formula: From periodogram based self-affinity parameter estimate (\code{sap}) to an informed estimate of the (fractal) dimension (FD).
}
\details{
The spectral slope will be converted to a dimension estimate using:

\deqn{D_{PSD}\approx\frac{3}{2}+\frac{14}{33}*\tanh\left(Slope * \ln(1+\sqrt{2})\right)}{D_{PSD} ≈ 3/2 + ((14/33)*tanh(sap*log(1+sqrt(2))))}
}
\examples{
# Informed FD of white noise
psd2fd(0)

# Informed FD of Brownian noise
psd2fd(-2)

# Informed FD of blue noise
psd2fd(2)
}
\author{
Fred Hasselman
}
\references{
Hasselman, F. (2013). When the blind curve is finite: dimension estimation and model inference based on empirical waveforms. Frontiers in Physiology, 4, 75. \url{http://doi.org/10.3389/fphys.2013.00075}
}

