% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCS-TB.R
\name{PSDslope}
\alias{PSDslope}
\title{PSDslope}
\usage{
PSDslope(y = ts(rnorm(n = 1024), frequency = 1), fs = frequency(y),
  nfft = 2^(nextpow2(length(y)/2)), fitRange = c(1, round(0.1 * nfft)),
  plot = FALSE)
}
\arguments{
\item{y}{A time series object, or a vector that can be converted to a time series object.}

\item{fs}{Sample frequency (defults to 1).}

\item{nfft}{Number of frequencies to estimate (defaults to next power of 2)}

\item{fitRange}{Vector of length 2 with range of frequencies to perform log-log fit.}

\item{plot}{Plot the log-log spectrum and slope.}
}
\description{
PSDslope
}

