% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdpsd.R
\name{sda2fd}
\alias{sda2fd}
\title{sda2fd}
\usage{
sda2fd(sap)
}
\arguments{
\item{sap}{Self-afinity parameter estimate based on SDA slope (e.g., \code{\link{fd.sda}})).}
}
\value{
An informed estimate of the Fractal Dimension, see Hasselman(2013) for details.
}
\description{
Conversion formula: Standardised Dispersion Analysis (SDA) estimate of self-affinity parameter (\code{sap}) to an informed estimate of the (fractal) dimension (FD).
}
\details{
Note that for some signals different PSD slope values project to a single SDA slope. That is, SDA cannot distinguish between all variaties of power-law scaling in the frequency domain.
}
\examples{
# Informed FD of white noise
sda2fd(-0.5)

# Informed FD of Brownian noise
sda2fd(-1)

# Informed FD of blue noise
sda2fd(-0.9)
}
\author{
Fred Hasselman
}
\references{
Hasselman, F. (2013). When the blind curve is finite: dimension estimation and model inference based on empirical waveforms. Frontiers in Physiology, 4, 75. \url{http://doi.org/10.3389/fphys.2013.00075}
}

